+++ src/Tail.php
@@ -6,9 +6,9 @@ use Eugenefvdm\Api\Contracts\TailInterface;
 
 class Tail extends Server implements TailInterface
 {
-    public function last(string $value, int $count = 1): array
+    public function last(string $filename, string $value, int $count = 1): array
     {
-        $command = "cat /var/log/mail.log | egrep -i '{$value}' | tail -n {$count}";
+        $command = "cat {$filename} | egrep -i '{$value}' | tail -n {$count}";
 
         return $this->executeCommand($command);
     }